# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI Receptionist API",
        "description": "AI-powered voice receptionist system for hospital appointment management. This API handles doctor schedules, patient information, appointment booking, and slot management. Some endpoints require API key authentication via the X-API-Key header (appointments and patient lookup).",
        "contact": {
            "name": "API Support",
            "email": "support@hospital.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Alternative Local Server"
        }
    ],
    "paths": {
        "/api/v1/appointments": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Book new appointment",
                "description": "Create a new appointment for a patient with a doctor",
                "operationId": "cd3cb3979bc60c0895c61ab950086885",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "doctor_id",
                                    "date",
                                    "start_time"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "description": "Patient ID",
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "doctor_id": {
                                        "description": "Doctor ID",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "date": {
                                        "description": "Appointment date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-09"
                                    },
                                    "start_time": {
                                        "description": "Start time in HH:MM format",
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00"
                                    },
                                    "slot_count": {
                                        "description": "Number of consecutive slots (default: 1)",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "type": {
                                        "description": "Appointment type (default: general)",
                                        "type": "string",
                                        "example": "general"
                                    },
                                    "reason": {
                                        "description": "Reason for appointment",
                                        "type": "string",
                                        "example": "Annual checkup"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment booked successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 8
                                                },
                                                "patient_id": {
                                                    "type": "integer",
                                                    "example": 7
                                                },
                                                "doctor_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-12-09"
                                                },
                                                "start_time": {
                                                    "type": "string",
                                                    "format": "time",
                                                    "example": "08:00:00"
                                                },
                                                "end_time": {
                                                    "type": "string",
                                                    "format": "time",
                                                    "example": "08:30:00"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "scheduled"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "general"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Booking failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/appointments/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get appointment details",
                "description": "Retrieve detailed information about a specific appointment",
                "operationId": "c9acb0dd9763645e9b48466da7120446",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Appointment ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 8
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 8
                                                },
                                                "patient": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 7
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Jane Doe"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "doctor": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 3
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "John Smith"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-12-09"
                                                },
                                                "start_time": {
                                                    "type": "string",
                                                    "format": "time",
                                                    "example": "08:00:00"
                                                },
                                                "end_time": {
                                                    "type": "string",
                                                    "format": "time",
                                                    "example": "08:30:00"
                                                },
                                                "slot_count": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "scheduled"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "general"
                                                },
                                                "reason": {
                                                    "type": "string",
                                                    "example": "Annual checkup"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Appointment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation error message"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key authentication. Use one of the configured API keys in the X-API-Key header.",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}